client/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brandybox"
version = "0.1.0"
description = "Brandy Box desktop client - sync local folder to Raspberry Pi storage"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Brandy Box" }]
dependencies = [
    "httpx>=0.27.0",
    "keyring>=25.0",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["brandybox*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
